+++ CHANGES.md
@@ -1,8 +1,12 @@
 # v0.2 -- UNRELEASED
 
 Don't catch exceptions in `dispatch`.
+
 Split the `dispatch` function into `dispatch` and `dispatch*`.
+
 Added `locate-commands`, for when invoking `dispatch*` directly.
+
+Detect command name conflicts when locating commands.
 
 # v0.1 -- 8 Aug 2022